requirements.txt
streamlit
pandas
numpy
tabulate
matplotlib
matplotlib
numpy
pandas
streamlit
tabulate
psycopg2-binary
